chore: Renames token color stylesheet
