+++ apps/site/components/Tokens/TokenColors.tsx
@@ -1,4 +1,4 @@
-import styles from './TokenColors.module.css'
+import styles from './token-colors.module.css'
 
 const TokenColors = ({ children }) => {
   return <ul className={styles.tokenColors}>{children}</ul>
